Cleaned up and reorganized code

+++ experimental/CoreBotWithClu/dotnet/13.core-bot-clu/Clu/CluEntity.cs
@@ -20,6 +20,6 @@ namespace Microsoft.BotBuilderSamples.Clu
         public int Length { get; set; }
 
         [JsonProperty("confidenceScore")]
-        public double ConfidenceScore { get; set; }
+        public float ConfidenceScore { get; set; }
     }
 }
